+++ inst/include/RcppArmadillo/Lighter
@@ -37,8 +37,8 @@
 #include <Rcpp/Lighter>
 
 // Remaining RcppArmadillo code
-#include <RcppArmadillo_impl/RcppArmadilloWrap.h>
-#include <RcppArmadillo_impl/RcppArmadilloAs.h>
-#include <RcppArmadillo_impl/RcppArmadilloSugar.h>
+#include <RcppArmadilloWrap.h>
+#include <RcppArmadilloAs.h>
+#include <RcppArmadilloSugar.h>
 
 #endif
