+++ inst/include/RcppArmadilloAs.h
