+++ inst/include/RcppArmadilloSugar.h
