+++ inst/include/RcppArmadilloWrap.h
